mv MetaTable to DataTable
