+++ tests/test_transform_meta.py
@@ -4,7 +4,7 @@ import pytest
 from pytest_cases import parametrize
 from sqlalchemy import Column, Integer
 
-from datapipe.metastore import MetaTable
+from datapipe.datatable import MetaTable
 from datapipe.step.batch_transform import BatchTransformStep
 from datapipe.store.database import DBConn
 from datapipe.types import MetaSchema
